move and export file path utils
