+++ packages/shared/src/index.ts
@@ -3,5 +3,6 @@ export * from './write-archive';
 export * from './utils/analytics';
 export * from './utils/logger';
 export * from './constants';
+export * from './utils/filePaths';
 
 export type { ChromaticConfig, ChromaticStorybookParameters } from './types';
