+++ packages/shared/src/utils/filePaths.ts
